Improve validation API docs (#990)
The `curl` uses `PUT` when using the `-T` or `--upload-file` flag, so
it's unnecessary to explicitly set the method.

The complex logic using temporary files and shell variables is also not
needed to determine if the validation failed. Using the `--fail` flag
has curl return with a non-zero exit code if the response has a 400 or
greater response code. The `--fail-with-body` prints the body, and in
this case it makes most sense, as we're interested in the details of the
validation failure.