Fix evaluation triggers for comment-only reviews (#1073)
GitHub reviews that use the "COMMENTED" state are processed both as
reviews and as regular comments. However, the handler for review events
ignored this and only considered the review state when determining if
the event affected approval. With certain policies, this meant that PRs
would only be approved after some other event triggered an evaluation.

Copy the relevant part of the logic from the issue comment handler in to
the review handler so that we can match comment-only reviews against the
normal comment patterns. Also clean up the handler to avoid an
unnecessary loop where a map lookup is sufficient.