Update Tailwind to v3 (#1076)
The biggest change here is the removal of purgecss in favor of the new
targeted generation. Otherwise, changes to configuration and existing
styles are relatively minor. 

While Tailwind v4 is the latest, I'm stopping at v3 because that's what
we're using in some other internal projects. It looks like v3 is still
supported by the upstream project (due to v4's change in browser
support), so this should resolve any dependency vulnerability issues. We
don't need any of the new features from v4 at this time.